Move upload_nightly
