+++ python/requirements.txt
@@ -17,11 +17,6 @@ ply == 3.8
 # For Cross-platform colored terminal text
 colorama == 0.3.7
 
-# For package uploading
-boto3 == 1.34.95
-pyOpenSSL == 24.0.0
-PyGithub == 1.58.1
-
 # For Python3 compatibility
 six == 1.16
